Fix: exported store
