+++ src/components/controls/index.tsx
@@ -3,11 +3,11 @@ import { X } from "lucide-react";
 import { useSelector } from "react-redux";
 import { twMerge } from "tailwind-merge";
 import { dataAttributes, ids } from "@/constants";
+import { Tool } from "@/constants";
 import { store } from "@/store";
 import { toggleControls } from "@/store/reducers/editor";
 import { ISTKProps } from "@/types";
 import { AnimatedSwitcher, IconButton } from "../core";
-import { Tool } from "../toolbar/data";
 import { default as ImageControls } from "./image";
 import { default as NoControls } from "./no-controls";
 import { default as NoSelectedElement } from "./no-selection";
