+++ src/components/workspace/dock/index.tsx
@@ -4,10 +4,10 @@ import { useSelector } from "react-redux";
 import { default as debounce } from "lodash/debounce";
 import { twMerge } from "tailwind-merge";
 import { dataAttributes, ids, selectors } from "@/constants";
+import { Tool } from "@/constants";
 import type { ISTKProps } from "@/types";
 import { d3Extended, getScaleFactorAccountingForViewBoxWidth } from "@/utils";
 import { Button } from "../../core";
-import { Tool } from "../../toolbar/data";
 import { showPostOffsetElements, showPreOffsetElements } from "../elements";
 import { default as Reload } from "../reload";
 import { default as DockHandler } from "./handler";
