+++ src/components/workspace/index.tsx
@@ -1,9 +1,9 @@
 import { useCallback } from "react";
 import { useSelector } from "react-redux";
 import { twMerge } from "tailwind-merge";
-import { dataAttributes, ids } from "@/constants";
+import { Tool, dataAttributes, ids } from "@/constants";
 import { type ISTKProps, SeatStatus } from "@/types";
-import { Tool, tools } from "../toolbar/data";
+import { tools } from "../toolbar/data";
 import { default as Crosshairs } from "./crosshairs";
 import { default as Dock } from "./dock";
 import { default as Element, ElementType } from "./elements";
