+++ src/stories/customization.mdx
@@ -16,7 +16,7 @@ return <SeatToolkit
   styles={{
     root: {
       className: "bg-gray-100",
-      style: {
+      properties: {
         padding: "20px",
         borderRadius: "10px",
         boxShadow: "0 0 10px rgba(0, 0, 0, 0.1)"
